Deprecate old QNode (#440)
* first attempt

* fix all imports

* fix more tests

* fix more tests

* fix optimization

* fix QNG

* fixing

* fix qg

* Fixing test_utils.py

* Change error type for tests in tensornet

* Adding model to the capabilities of in vqe test

* Fix AutogradQnode queue tests

* Refactor VQE for torch and tf (QNode already creates specific qnode objects), correct test_vqe

* Fix numpy array test case for parameter empty check, fix test messages

* Correct docs

* Remove trailing whitespace

* attempted docs fixes

* fix docs

* update pip version

* remove redundant theme info

* fix incorrect VQE docstring

* Revert to_tf() change

* Update tests/test_quantum_gradients.py


* Remove recorder return statements in test_templates_layers.py

* Remove recorder return statements in test_templates_subroutines.py

* Changelog

* Add comment for numpy.array check
