+++ .github/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Breaking changes
 
+* Deprecated the old `QNode` such that only the new `QNode` and its syntax can be used,
+  moved all related files from the `pennylane/beta` folder to `pennylane`.
+  [#440](https://github.com/XanaduAI/pennylane/pull/440)
+
 ### Improvements
 
 ### Documentation
@@ -19,7 +23,7 @@
 
 This release contains contributions from (in alphabetical order):
 
-Ville Bergholm
+Ville Bergholm, Josh Izaac, Antal Száva
 
 ---
