+++ doc/code/qml_beta.rst
@@ -6,16 +6,9 @@ and features.
 
 .. currentmodule:: pennylane.beta
 
-.. automodapi:: pennylane.beta.interfaces
-    :include-all-objects:
-    :no-inheritance-diagram:
-
 .. automodapi:: pennylane.beta.plugins
     :include-all-objects:
     :no-inheritance-diagram:
-
-.. automodapi:: pennylane.beta.qnodes
-    :include-all-objects:
 
 .. automodapi:: pennylane.beta.vqe
     :include-all-objects:
