+++ doc/code/qml_operation.rst
@@ -14,4 +14,4 @@ qml.operation
 .. automodapi:: pennylane.operation
     :no-heading:
     :include-all-objects:
-    :skip: QNode, Sequence, Enum, IntEnum, Variable, ClassPropertyDescriptor
+    :skip: Sequence, Enum, IntEnum, Variable, ClassPropertyDescriptor
