+++ doc/code/qml_qnodes.rst
@@ -1,0 +1,16 @@
+qml.qnodes
+==========
+
+.. warning::
+
+    Unless you are a PennyLane or plugin developer, you likely do not need
+    to use these classes directly.
+
+    See the :doc:`quantum circuits <../introduction/circuits>` page for more
+    details on creating QNodes, as well as the :func:`~.qnode` decorator
+    and :func:`~.QNode` constructor.
+
+.. automodapi:: pennylane.qnodes
+    :no-heading:
+    :include-all-objects:
+    :skip: qnode, QNode, QuantumFunctionError
