+++ doc/index.rst
@@ -194,6 +194,7 @@ PennyLane is **free** and **open source**, released under the Apache License, Ve
    code/qml_interfaces
    code/qml_operation
    code/qml_plugins
+   code/qml_qnodes
    code/qml_templates
    code/qml_utils
    code/qml_variable
