+++ doc/xanadu_theme/layout.html
@@ -2,9 +2,6 @@
 
 {# Do this so that bootstrap is included before the main css file #}
 {%- block htmltitle %}
-  {% set script_files = script_files + ["_static/myscript.js"] %}
-
-  {% set script_files = script_files + ["_static/myscript.js"] %}
   <link href="https://fonts.googleapis.com/css?family=Noto+Serif" rel="stylesheet">
   <link href="https://fonts.googleapis.com/css?family=Roboto&display=swap" rel="stylesheet">
