+++ tests/qnodes/test_qnode_cv.py
@@ -20,8 +20,8 @@ import numpy as np
 import pennylane as qml
 from pennylane._device import Device
 from pennylane.operation import CVObservable
-from pennylane.beta.qnodes.base import QuantumFunctionError
-from pennylane.beta.qnodes.cv import CVQNode
+from pennylane.qnodes.base import QuantumFunctionError
+from pennylane.qnodes.cv import CVQNode
 
 
 class PolyN(qml.ops.PolyXP):
